% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inAdmin.R
\docType{data}
\name{inAdmin}
\alias{inAdmin}
\title{A sample administraive boundaries dataset describing admin2 boundaries in Pakistan}
\format{a SpatialPolygonsDataFrame of the admin2 Pakistan boundaries}
\usage{
data(inAdmin)
}
\description{
To load the sample data, data(inAdmin)
}
\keyword{datasets}
