% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inPts.R
\docType{data}
\name{inPts}
\alias{inPts}
\title{A sample points dataset for testing point jittering}
\format{a SpatialPointsDataFrame of randomly dropped point locations inside the Pakistan admin2 boundaries}
\usage{
data(inPts)
}
\description{
To load the sample data, data(inPts)
}
\keyword{datasets}
